+++ src/utils/openai-format.ts
